New archive packing/unpacking implementation
The motivation for this new implementation is the memory
crashes seen in production. These occur when merging archives.
The original archive implementation held a lot of data in
memory when merging.

The new implemenation is still in-memory, but is much more
memory efficient and written in a way that using disk space
for merging instead of memory will be very easy. For the moment,
we will see how it performs in the production.

As well as being more efficient, the new implementation is
dramatically simpler. The packing/unpacking process is contained
behind a very simple API in the archive package.